backend/requirements/dev.txt
-r base.txt

# ===========================
# Development Tools
# ===========================
django-debug-toolbar>=4.4,<5.0
django-extensions>=3.2,<4.0
ipython>=8.0,<9.0

# ===========================
# Testing
# ===========================
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
faker>=28.0,<30.0

# ===========================
# Code Quality
# ===========================
ruff>=0.6,<1.0
black>=24.0,<25.0
isort>=5.13,<6.0
mypy>=1.11,<2.0
django-stubs>=5.0,<6.0
djangorestframework-stubs>=3.15,<4.0
